Handling decoupling of CustomSearchEngine controller and its view and folders
